+++ packages/stripe_js/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 7.5.0
+
+ - stability improvements
+
+## 7.4.0
+**Features**
+- Add WASM support
+
 ## 7.3.0
  - Small fixes and maitenance improvements