+++ packages/stripe_web/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 7.6.0
+- stability improvements
+
+## 7.5.0
+- Update bindings
+
+## 7.4.0
+**Features**
+- Add WASM support
+
 ## 7.3.0
  - Small fixes and maitenance improvements